+++ test/tls/test_ciphers.py
@@ -95,13 +95,26 @@ class TestTlsCiphers(TestTlsCases.TestTlsBase):
 
     @mock.patch.object(
         AnalyzerCipherSuites, '_next_accepted_cipher_suites',
-        side_effect=TlsAlert(TlsAlertDescription.PROTOCOL_VERSION)
+        side_effect=StopIteration,
     )
     @mock.patch.object(
         AnalyzerCipherSuites, '_get_accepted_cipher_suites_fallback',
         return_value=[]
     )
     def test_error_protocol_version(self, mocked_next_accepted_cipher_suites, _):
+        result = self.get_result('rc4.badssl.com', 443)
+        self.assertEqual(len(result.cipher_suites), 0)
+        self.assertEqual(mocked_next_accepted_cipher_suites.call_count, 1)
+
+    @mock.patch.object(
+        AnalyzerCipherSuites, '_next_accepted_cipher_suites',
+        side_effect=TlsAlert(TlsAlertDescription.DECODE_ERROR),
+    )
+    @mock.patch.object(
+        AnalyzerCipherSuites, '_get_accepted_cipher_suites_fallback',
+        return_value=[]
+    )
+    def test_error_decode_error(self, mocked_next_accepted_cipher_suites, _):
         result = self.get_result('rc4.badssl.com', 443)
         self.assertEqual(len(result.cipher_suites), 0)
         self.assertEqual(mocked_next_accepted_cipher_suites.call_count, 1)